Changes to logic with new custom exceptions

+++ src/main/java/bucholc/arkadiusz/taskManagementApp/service/UserService.java
@@ -1,7 +1,10 @@
 package bucholc.arkadiusz.taskManagementApp.service;
 
+import bucholc.arkadiusz.taskManagementApp.exception.UserAlreadyExistsException;
+import bucholc.arkadiusz.taskManagementApp.exception.UserAssignedToTaskException;
 import bucholc.arkadiusz.taskManagementApp.exception.UserNotFoundException;
 import bucholc.arkadiusz.taskManagementApp.model.User;
+import bucholc.arkadiusz.taskManagementApp.repository.TaskRepository;
 import bucholc.arkadiusz.taskManagementApp.repository.UserRepository;
 import org.springframework.stereotype.Service;
 
@@ -11,20 +14,30 @@ import java.util.Map;
 @Service
 public class UserService {
 	private final UserRepository userRepository;
+	private final TaskRepository taskRepository;
 
-	public UserService(UserRepository userRepository) {
+	public UserService(UserRepository userRepository, TaskRepository taskRepository) {
 		this.userRepository = userRepository;
+		this.taskRepository = taskRepository;
 	}
 	
 	public List<User> findAll() {
 		return userRepository.findAll();
 	}
-	
-	public User saveUser(User user) {
+
+	public User createUser(User user) {
+		if (userRepository.existsByEmail(user.getEmail())) {
+			throw new UserAlreadyExistsException("User with email " + user.getEmail() + " already exists");
+		}
+		
 		return userRepository.save(user);
 	}
 	
 	public void removeUser(Long id) {
+		boolean isUserAssignedToTask = taskRepository.existsByAssignedUsersId(id);
+		if (isUserAssignedToTask) {
+			throw new UserAssignedToTaskException("User with id " + id + " assigned to task, cannot be removed");
+		}
 		userRepository.deleteById(id);
 	}
 	
@@ -32,7 +45,7 @@ public class UserService {
 		return userRepository.findById(id).orElse(null);
 	}
 	
-	public User partialUpdate(Long id, Map<String, Object> updatesMap) {
+	public User updateUser(Long id, Map<String, Object> updatesMap) {
 		User user = userRepository.findById(id).orElseThrow(() -> new UserNotFoundException(id));
 		updatesMap.forEach((key, value) -> {
 			switch (key) {
@@ -45,6 +58,8 @@ public class UserService {
 				case "email":
 					user.setEmail((String) value);
 					break;
+			default:
+				throw new IllegalStateException("Unexpected value: " + key);
 			}
 		});
 		return userRepository.save(user);
